requirements.txt
# コア
streamlit==1.28.0
openai==1.56.0
pandas==2.1.0
python-dotenv==1.0.0

# ファイル処理
python-docx==1.1.0
PyPDF2==3.0.1
chardet==5.2.0

# その他
openpyxl==3.1.2
